techDebt(buRefactorTraining): Move uncheckedEstablishment functionality into file in new shared folder
