+++ server/routes/establishments/bulkUpload/validate/workers/validateWorkers.js
@@ -2,6 +2,7 @@ const { validateWorkerCsv } = require('./validateWorkerCsv');
 const { validateDuplicateWorkerID } = require('./validateDuplicateWorkerID');
 const { validatePartTimeSalaryNotEqualToFTE } = require('./validatePartTimeSalaryNotEqualToFTE');
 const { validateWorkerUnderNationalInsuranceMaximum } = require('./validateWorkerUnderNationalInsuranceMaximum');
+const { establishmentNotFoundInFile, addNoEstablishmentError } = require('../shared/uncheckedEstablishment');
 
 const validateWorkers = async (workers, myCurrentEstablishments, allEstablishmentsByKey, myAPIEstablishments) => {
   const workersKeyed = [];
@@ -41,7 +42,7 @@ const validateWorkers = async (workers, myCurrentEstablishments, allEstablishmen
     const establishmentKey = thisWorker.localId ? thisWorker.localId.replace(/\s/g, '') : '';
 
     if (establishmentNotFoundInFile(allEstablishmentsByKey, establishmentKey)) {
-      addNoEstablishmentError(csvWorkerSchemaErrors, thisWorker);
+      addNoEstablishmentError(csvWorkerSchemaErrors, thisWorker, 'Worker');
       deleteWorker(myAPIWorkers, thisWorker.lineNumber);
       return;
     }
@@ -79,35 +80,11 @@ const createWorkerKey = (localEstablishmentId, workerId) => {
   return (localEstablishmentId + workerId).replace(/\s/g, '');
 };
 
-const uncheckedEstablishment = (thisWorker) => {
-  return {
-    origin: 'Workers',
-    lineNumber: thisWorker.lineNumber,
-    errCode: UNCHECKED_ESTABLISHMENT_ERROR(),
-    errType: 'UNCHECKED_ESTABLISHMENT_ERROR',
-    error: 'LOCALESTID does not exist in Workplace file',
-    source: thisWorker.localId,
-    column: 'LOCALESTID',
-    worker: thisWorker.uniqueWorkerId,
-    name: thisWorker.localId,
-  };
-};
-
-const UNCHECKED_ESTABLISHMENT_ERROR = () => 997;
-
-const establishmentNotFoundInFile = (allEstablishmentsByKey, establishmentKey) =>
-  !allEstablishmentsByKey[establishmentKey];
-
-const addNoEstablishmentError = (csvWorkerSchemaErrors, thisWorker) =>
-  csvWorkerSchemaErrors.push(uncheckedEstablishment(thisWorker));
-
 const deleteWorker = (myAPIWorkers, workerLineNumber) => delete myAPIWorkers[workerLineNumber];
 
 module.exports = {
   validateWorkers,
   createKeysForWorkers,
   createWorkerKey,
-  establishmentNotFoundInFile,
-  addNoEstablishmentError,
   deleteWorker,
 };
